+++ src/compose/images.ts
@@ -24,38 +24,11 @@ import { strict as assert } from 'assert';
 import log from '../lib/supervisor-console';
 import { setTimeout } from 'timers/promises';
 
+import type { Image } from './types';
+export type { Image } from './types';
+
 interface FetchProgressEvent {
 	percentage: number;
-}
-
-export interface Image {
-	id?: number;
-	/**
-	 * image [registry/]repo@digest or [registry/]repo:tag
-	 */
-	name: string;
-	/**
-	 * @deprecated to be removed in target state v4
-	 */
-	appId: number;
-	appUuid: string;
-	/**
-	 * @deprecated to be removed in target state v4
-	 */
-	serviceId: number;
-	serviceName: string;
-	/**
-	 * @deprecated to be removed in target state v4
-	 */
-	imageId: number;
-	/**
-	 * @deprecated to be removed in target state v4
-	 */
-	releaseId: number;
-	commit: string;
-	dockerImageId?: string;
-	status?: 'Downloading' | 'Downloaded' | 'Deleting';
-	downloadProgress?: number | null;
 }
 
 // Setup an event emitter
