+++ src/compose/types/application-manager.ts
@@ -1,0 +1,3 @@
+import type { App } from './app';
+
+export type InstancedAppState = { [appId: number]: App };
