+++ src/compose/types/index.ts
@@ -1,2 +1,7 @@
 export * from './service';
 export * from './network';
+export type * from './volume';
+export type * from './image';
+export type * from './composition-step';
+export type * from './app';
+export type * from './application-manager';
