+++ src/compose/types/network.ts
@@ -1,4 +1,7 @@
-import type { NetworkInspectInfo as DockerNetworkInspectInfo } from 'dockerode';
+import type {
+	NetworkInspectInfo as DockerNetworkInspectInfo,
+	NetworkCreateOptions,
+} from 'dockerode';
 
 // TODO: ConfigOnly is part of @types/dockerode@v3.2.0, but that version isn't
 // compatible with `resin-docker-build` which is used for `npm run sync`.
@@ -43,4 +46,19 @@ export interface NetworkConfig {
 	labels: { [labelName: string]: string };
 	options: { [optName: string]: string };
 	configOnly: boolean;
+}
+
+export interface Network {
+	appId: number;
+	appUuid?: string;
+	name: string;
+	config: NetworkConfig;
+
+	isEqualConfig(network: Network): boolean;
+	create(): Promise<void>;
+	remove(): Promise<void>;
+	toDockerConfig(): NetworkCreateOptions & {
+		ConfigOnly: boolean;
+	};
+	toComposeObject(): ComposeNetworkConfig;
 }
