+++ test/integration/compose/application-manager.spec.ts
@@ -19,7 +19,7 @@ import {
 	expectSteps,
 	expectNoStep,
 } from '~/test-lib/state-helper';
-import type { InstancedAppState } from '~/src/types';
+import type { InstancedAppState } from '~/src/compose/types';
 
 // TODO: application manager inferNextSteps still queries some stuff from
 // the engine instead of receiving that information as parameter. Refactoring
